Traitement espace commentaire , gestion des rôles coté admin

+++ src/Controller/BlogpostController.php
@@ -3,8 +3,7 @@
 namespace App\Controller;
 
 use App\Model\BlogPostManager;
-use Twig\Extra\String\StringExtension;
-
+use App\Model\CommentManager;
 
 
 class BlogpostController extends AbstractController
@@ -16,15 +15,21 @@ class BlogpostController extends AbstractController
     public function index(): string
     {
         //si connecté
-        if($_SESSION['is_connected']===true ){
-
+        if (isset($_SESSION) && $_SESSION['is_connected'] === true && $_SESSION['admin'] == 2){
             $blogpostManager = new BlogPostManager();
             $blogs = $blogpostManager->selectAll();
             return $this->twig->render('blogpost/index.html.twig', [
-                'blogs' => $blogs
+                'blogs' => $blogs,
             ]);
         }
         //
+       /* $blogpostManager = new BlogPostManager();
+        $blogs = $blogpostManager->selectAll();
+        return $this->twig->render('blogpost/index.html.twig', [
+            'blogs' => $blogs,
+        ]);
+        //
+       */
     }
 
 
@@ -33,12 +38,21 @@ class BlogpostController extends AbstractController
      */
     public function show(int $id): string
     {
-        $blogpostManager = new BlogPostManager();
-        $blogpost = $blogpostManager->selectOneById($id);
+        //si connecté et le role est admin
+        if (isset($_SESSION) && $_SESSION['is_connected'] === true && $_SESSION['admin'] == 2) {
+            $blogpostManager = new BlogPostManager();
+            $blogpost = $blogpostManager->selectOneById($id);
 
-        return $this->twig->render('blogpost/show.html.twig', [
-            'blogpost' => $blogpost
-        ]);
+            //test commentaire
+            $commentManager = new CommentManager();
+            $comments = $commentManager->selectAll();
+            //test commentaire
+
+            return $this->twig->render('blogpost/show.html.twig', [
+                'blogpost' => $blogpost,
+                'comments' => $comments
+            ]);
+        }//
     }
 
 
@@ -47,22 +61,25 @@ class BlogpostController extends AbstractController
      */
     public function edit(int $id): string
     {
-        $blogpostManager = new BlogPostManager();
-        $blogpost = $blogpostManager->selectOneById($id);
-        if ($_SERVER['REQUEST_METHOD'] === 'POST') {
-            // clean $_POST data
-            $blogpost = array_map('trim', $_POST);
+        //si connecté et le role est admin
+        if (isset($_SESSION) && $_SESSION['is_connected'] === true && $_SESSION['admin'] == 2) {
+            $blogpostManager = new BlogPostManager();
+            $blogpost = $blogpostManager->selectOneById($id);
+            if ($_SERVER['REQUEST_METHOD'] === 'POST') {
+                // clean $_POST data
+                $blogpost = array_map('trim', $_POST);
 
-            // TODO validations (length, format...)
+                // TODO validations (length, format...)
 
-            // if validation is ok, update and redirection
-            $blogpostManager->update($blogpost);
-            header('Location: /blogpost/show/' . $id);
-        }
+                // if validation is ok, update and redirection
+                $blogpostManager->update($blogpost);
+                header('Location: /blogpost/show/' . $id);
+            }
 
-        return $this->twig->render('blogpost/edit.html.twig', [
-            'blogpost' => $blogpost,
-        ]);
+            return $this->twig->render('blogpost/edit.html.twig', [
+                'blogpost' => $blogpost,
+            ]);
+        }//
     }
 
 
@@ -71,19 +88,22 @@ class BlogpostController extends AbstractController
      */
     public function add(): string
     {
-        if ($_SERVER['REQUEST_METHOD'] === 'POST') {
-            // clean $_POST data
-            $blogpost = array_map('trim', $_POST);
+        //si connecté et le rôle est admin
+        if (isset($_SESSION) && $_SESSION['is_connected'] === true && $_SESSION['admin'] == 2) {
+            if ($_SERVER['REQUEST_METHOD'] === 'POST') {
+                // clean $_POST data
+                $blogpost = array_map('trim', $_POST);
 
-            // TODO validations (length, format...)
+                // TODO validations (length, format...)
 
-            // if validation is ok, insert and redirection
-            $blogpostManager = new BlogPostManager();
-            $id = $blogpostManager->insert($blogpost);
-            header('Location:/blogpost/show/' . $id);
-        }
+                // if validation is ok, insert and redirection
+                $blogpostManager = new BlogPostManager();
+                $id = $blogpostManager->insert($blogpost);
+                header('Location:/blogpost/show/' . $id);
+            }
 
-        return $this->twig->render('blogpost/add.html.twig');
+            return $this->twig->render('blogpost/add.html.twig');
+        }//
     }
 
 
@@ -92,10 +112,19 @@ class BlogpostController extends AbstractController
      */
     public function delete(int $id)
     {
-        if ($_SERVER['REQUEST_METHOD'] === 'POST') {
-            $blogpostManager = new BlogPostManager();
-            $blogpostManager->delete($id);
-            header('Location:/blogpost/index');
-        }
+        //si connecté et rôle est admin
+        if (isset($_SESSION) && $_SESSION['is_connected'] === true && $_SESSION['admin'] == 2) {
+            if ($_SERVER['REQUEST_METHOD'] === 'POST') {
+                $blogpostManager = new BlogPostManager();
+                $blogpostManager->delete($id);
+                header('Location:/blogpost/index');
+            }
+        }//
     }
+
+
+
+    //ICI
+
+    //
 }
